more factory cleanup, hopeful v6 fix
