+++ source/common/http/http3/BUILD
@@ -13,6 +13,7 @@ envoy_cc_library(
     srcs = ["conn_pool.cc"],
     hdrs = ["conn_pool.h"],
     deps = [
+        ":quic_client_connection_factory_lib",
         "//include/envoy/event:dispatcher_interface",
         "//include/envoy/upstream:upstream_interface",
         "//source/common/http:codec_client_lib",
@@ -27,6 +28,15 @@ envoy_cc_library(
     deps = [
         "//include/envoy/config:typed_config_interface",
         "//include/envoy/http:codec_interface",
+        "//include/envoy/network:connection_interface",
+    ],
+)
+
+envoy_cc_library(
+    name = "quic_client_connection_factory_lib",
+    hdrs = ["quic_client_connection_factory.h"],
+    deps = [
+        "//include/envoy/config:typed_config_interface",
         "//include/envoy/network:connection_interface",
         "//include/envoy/ssl:context_config_interface",
         "@envoy_api//envoy/config/listener/v3:pkg_cc_proto",
