+++ source/common/network/utility.cc
@@ -538,12 +538,6 @@ Api::IoCallUint64Result Utility::writeToSocket(IoHandle& handle, Buffer::RawSlic
   if (send_result.ok()) {
     ENVOY_LOG_MISC(trace, "sendmsg bytes {}", send_result.rc_);
   } else {
-    // TODO(alyssawilk) REMOVE BEFORE SUBMITTING.
-    std::cerr << "Failed sending. Peer address " << peer_address.asString();
-    if (local_ip) {
-      std::cerr << " local ip " << local_ip->addressAsString();
-    }
-    std::cerr << std::endl;
     ENVOY_LOG_MISC(debug, "sendmsg failed with error code {}: {}",
                    static_cast<int>(send_result.err_->getErrorCode()),
                    send_result.err_->getErrorDetails());
