+++ source/extensions/quic_listeners/quiche/BUILD
@@ -136,9 +136,9 @@ envoy_cc_library(
 )
 
 envoy_cc_library(
-    name = "codec_lib",
-    srcs = ["codec_impl.cc"],
-    hdrs = ["codec_impl.h"],
+    name = "client_connection_factory_lib",
+    srcs = ["client_connection_factory_impl.cc"],
+    hdrs = ["client_connection_factory_impl.h"],
     tags = ["nofips"],
     deps = [
         ":active_quic_listener_lib",
@@ -150,9 +150,27 @@ envoy_cc_library(
         ":envoy_quic_utils_lib",
         "//include/envoy/http:codec_interface",
         "//include/envoy/registry",
+        "//source/common/http/http3:quic_client_connection_factory_lib",
         "//source/common/http/http3:quic_codec_factory_lib",
         "//source/common/http/http3:well_known_names",
         "//source/extensions/transport_sockets/tls:ssl_socket_lib",
+        "@com_googlesource_quiche//:quic_core_http_spdy_session_lib",
+    ],
+)
+
+envoy_cc_library(
+    name = "codec_lib",
+    srcs = ["codec_impl.cc"],
+    hdrs = ["codec_impl.h"],
+    tags = ["nofips"],
+    deps = [
+        ":envoy_quic_client_session_lib",
+        ":envoy_quic_server_session_lib",
+        ":envoy_quic_utils_lib",
+        "//include/envoy/http:codec_interface",
+        "//include/envoy/registry",
+        "//source/common/http/http3:quic_codec_factory_lib",
+        "//source/common/http/http3:well_known_names",
         "@com_googlesource_quiche//:quic_core_http_spdy_session_lib",
     ],
 )
